Move music.EncoderPipeline to pipelines package
