+++ magenta/music/__init__.py
@@ -47,7 +47,6 @@ from magenta.music.drums_lib import extract_drum_tracks
 from magenta.music.drums_lib import midi_file_to_drum_track
 
 from magenta.music.encoder_decoder import ConditionalEventSequenceEncoderDecoder
-from magenta.music.encoder_decoder import EncoderPipeline
 from magenta.music.encoder_decoder import EventSequenceEncoderDecoder
 from magenta.music.encoder_decoder import LookbackEventSequenceEncoderDecoder
 from magenta.music.encoder_decoder import MultipleEventSequenceEncoder
