+++ magenta/music/encoder_decoder.py
@@ -52,7 +52,6 @@ import abc
 import numbers
 
 from magenta.music import constants
-from magenta.pipelines import pipeline
 import numpy as np
 from six.moves import range  # pylint: disable=redefined-builtin
 import tensorflow as tf
@@ -1009,28 +1008,6 @@ class MultipleEventSequenceEncoder(EventSequenceEncoderDecoder):
 
   def class_index_to_event(self, class_index, events):
     raise NotImplementedError
-
-
-class EncoderPipeline(pipeline.Pipeline):
-  """A pipeline that converts an EventSequence to a model encoding."""
-
-  def __init__(self, input_type, encoder_decoder, name=None):
-    """Constructs an EncoderPipeline.
-
-    Args:
-      input_type: The type this pipeline expects as input.
-      encoder_decoder: An EventSequenceEncoderDecoder.
-      name: A unique pipeline name.
-    """
-    super(EncoderPipeline, self).__init__(
-        input_type=input_type,
-        output_type=tf.train.SequenceExample,
-        name=name)
-    self._encoder_decoder = encoder_decoder
-
-  def transform(self, seq):
-    encoded = self._encoder_decoder.encode(seq)
-    return [encoded]
 
 
 def make_sequence_example(inputs, labels):
